+++ src/layman/layer/__init__.py
@@ -68,7 +68,7 @@ PUBLICATION_TYPES = {
             ('layman.layer.prime_db_schema.table', InternalSourceTypeDef(info_items=[
                 'access_rights', 'name', 'workspace', 'title', 'uuid', 'bounding_box', 'style_type', 'native_crs',
                 'native_bounding_box', 'geodata_type', 'updated_at', 'id', 'type', 'image_mosaic', 'table_uri',
-                'original_data_source', 'wfs_wms_status', 'used_in_maps', 'description', 'created_at', 'is_public_workspace', ]),),
+                'original_data_source', 'wfs_wms_status', 'used_in_maps', 'description', 'created_at', 'is_public_workspace', 'file_path', ]),),
             ('layman.layer.filesystem.input_chunk', InternalSourceTypeDef(info_items=['file', ]),),
             ('layman.layer.filesystem.input_file', InternalSourceTypeDef(info_items=['file', ]),),
             ('layman.layer.filesystem.input_style', InternalSourceTypeDef(info_items=[]),),
@@ -128,7 +128,7 @@ PUBLICATION_TYPES = {
             settings.GEODATA_TYPE_RASTER: {
                 'name', 'uuid', 'layman_metadata', 'url', 'title', 'description', 'updated_at', 'wms', 'thumbnail', 'file', 'metadata',
                 'style', 'access_rights', 'bounding_box', 'native_crs', 'native_bounding_box', 'image_mosaic',
-                'original_data_source', 'geodata_type', 'used_in_maps',
+                'original_data_source', 'geodata_type', 'used_in_maps', 'file_path',
             },
             settings.GEODATA_TYPE_UNKNOWN: {
                 'name', 'uuid', 'layman_metadata', 'url', 'title', 'description', 'updated_at', 'wms', 'thumbnail', 'file', 'metadata',